+++ Routes/iot.js
@@ -8,7 +8,6 @@ module.exports = function (app, db) {
     const iot_registerItem = require('./iot_registerItem');
     const iot_editItem = require('./iot_editItem');
     const iot_help = require('./iot_help');
-    const iot_mypage = require('./iot_mypage');
 
     var check = (req, res, next) => {
         var id = req.session['memberID'];
@@ -27,7 +26,7 @@ module.exports = function (app, db) {
 
     router.get('/', (req, res, next) => { iot_monitoring.init(req, res, db) });
 
-    router.post('/', (req, res, next) => { iot_mypage.sessionCheck(req, res, db) });
+    router.post('/', (req, res, next) => { iot_monitoring.sessionCheck(req, res, db) });
 
     router.get('/monitoring', (req, res, next) => { iot_monitoring.init(req, res, db) });
 
@@ -37,8 +36,6 @@ module.exports = function (app, db) {
         itemlist = JSON.parse(itemlist);
         res.render('Iot/warehousing', {'itemlist': itemlist, 'userType': userType});
     });
-
-    router.get('/statistics', (req, res, next) => { iot_statistics.init(req, res, db) });
 
     router.get('/help', (req, res, next) => { iot_help.init(req, res, db) });
