+++ Views/User/Admin/ad_RequestIoT.ejs
@@ -2,15 +2,6 @@
 <html lang="en" dir="ltr">
 <head>
   <%- include('../../head') %>
-  <title></title>
-  <style media="screen">
-    .card-header {
-      font-size: 25px
-    }
-    .page-link {
-      font-size: 20px !important
-    }
-  </style>
 </head>
 
 <body>
